refactor: TravelAddViewController에 loadView를 사용해 UI를 불러오도록 구현
